requirements.txt
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==2.2.4
django-bootstrap4==0.0.8
django-heroku==0.3.1
django-model-utils==3.2.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
whitenoise==4.1.3
